project/server/requirements.txt
fastapi
uvicorn
httpx
motor
pymongo
pydantic
pydantic[email]
python-dotenv
